Move Images functionality to new files (ImageCommands.ts and ImageService.ts)
