+++ ts/commands/ElementCommands.ts
@@ -63,42 +63,6 @@ class ElementCommands {
         const config = configService.mergeConfig(options);
         elementService.stopMaintenanceMode(args.id, config, logger);
       });
-
-    (caporal as any)
-      .command('image custom delete', 'Delete a custom image')
-      .option('--username', 'Metricly Username')
-      .option('--password', 'Metricly Password')
-      .option('--profile', 'Metricly profile', /.*/, 'default')
-      .option('--format', 'output format: text, json', ['text', 'json'], 'text')
-      .argument('<name>', 'Image Name')
-      .action((args, options, logger) => {
-        const config = configService.mergeConfig(options);
-        elementService.customtypeImageRm(args.name, config, logger);
-      });
-
-    (caporal as any)
-      .command('image custom set', 'Set images for custom element types')
-      .option('--username', 'Metricly Username')
-      .option('--password', 'Metricly Password')
-      .option('--profile', 'Metricly profile', /.*/, 'default')
-      .option('--format', 'output format: text, json', ['text', 'json'], 'text')
-      .argument('<name>', 'Image Name')
-      .argument('<file>', 'Image File Name')
-      .action((args, options, logger) => {
-        const config = configService.mergeConfig(options);
-        elementService.customtypeImageSet(args.name, args.file, config, logger);
-      });
-
-    (caporal as any)
-      .command('image custom list', 'List images for custom element types')
-      .option('--username', 'Metricly Username')
-      .option('--password', 'Metricly Password')
-      .option('--profile', 'Metricly profile', /.*/, 'default')
-      .option('--format', 'output format: text, json', ['text', 'json'], 'text')
-      .action((args, options, logger) => {
-        const config = configService.mergeConfig(options);
-        elementService.customtypeImageList(config, logger);
-      });
   }
 }
